refactor: briefStationInfoWindowStore 모듈화 (#92)
[#86]
