+++ frontend/src/components/google-maps/map/CarFfeineMap.tsx
@@ -1,5 +1,4 @@
 import { useEffect, useRef } from 'react';
-import { createRoot } from 'react-dom/client';
 
 import { useCurrentPosition } from '../../../hooks/useCurrentPosition';
 import UserMarker from '../marker/UserMarker';
@@ -8,7 +7,6 @@ import { useUpdateStations } from '../../../hooks/useUpdateStations';
 import { useExternalState } from '../../../utils/external-state';
 import { googleMapStore } from '../../../stores/googleMapStore';
 import StationList from '../../ui/StationList';
-import { briefStationInfoWindowStore } from '../../../stores/briefStationInfoWindowStore';
 import MarkerList from '../../ui/MarkerList';
 import { INITIAL_ZOOM_SIZE } from '../../../constants';
 
@@ -42,10 +40,10 @@ const CarFfeineMap = () => {
       {isClientReady && (
         <>
           <CarFfeineMapListener googleMap={googleMap} />
-          <CarFfeineInfoWindowInitializer />
-
           <StationMarkersContainer googleMap={googleMap} />
           <UserMarker googleMap={googleMap} position={position} />
+          <StationList />
+          <MarkerList />
         </>
       )}
     </>
@@ -73,38 +71,6 @@ const CarFfeineMapListener = ({ googleMap }: Props) => {
   }, []);
 
   return <></>;
-};
-
-const CarFfeineInfoWindowInitializer = () => {
-  const [briefStationInfoWindow, setBriefStationInfoWindow] = useExternalState(
-    briefStationInfoWindowStore
-  );
-
-  useEffect(() => {
-    const container = document.createElement('div');
-    const briefStationInfoRoot = createRoot(container);
-    const infoWindowInstance = new google.maps.InfoWindow({
-      content: container,
-    });
-
-    const initialBriefStationInfoWindow = {
-      briefStationInfoRoot,
-      infoWindowInstance,
-    };
-
-    setBriefStationInfoWindow(initialBriefStationInfoWindow);
-  }, []);
-
-  return (
-    <>
-      {briefStationInfoWindow && (
-        <>
-          <StationList />
-          <MarkerList />
-        </>
-      )}
-    </>
-  );
 };
 
 export default CarFfeineMap;
